# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de base en laravel",
        "description": "MDN",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://mdn-rdv.com/api"
        }
    ],
    "paths": {
...
        "/admin/admins/store": {
            "post": {
                "tags": [
                    "ADMINISTRATION__ADMINISTRATEURS"
                ],
                "summary": "Enregistre un administrateur",
                "description": "",
                "operationId": "23c218234ac81ba86f01eeb316c6a13a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Mulamba"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "bgi@gmail.com"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "0852277379"
                                    },
                                    "role": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/admins/show/{id}": {
            "get": {
                "tags": [
                    "ADMINISTRATION__ADMINISTRATEURS"
                ],
                "summary": "Renvoie les informations d'un administrateur par son ID",
                "operationId": "f2177002a53a5274642106a32a289fcf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 18
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/admins/update": {
            "post": {
                "tags": [
                    "ADMINISTRATION__ADMINISTRATEURS"
                ],
                "summary": "Editer les informations d'un administrateur",
                "operationId": "f23917d287b53b769e2ed799bcb1fc63",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Mulamba"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "bgi@gmail.com"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "0852277379"
                                    },
                                    "role": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/admins/activate/{id}": {
            "get": {
                "tags": [
                    "ADMINISTRATION__ADMINISTRATEURS"
                ],
                "summary": "Active ou désactive l'état d'un administrateur par son ID",
                "operationId": "ca17de4ed94ef4df58c48750e7d5afea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 18
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/admins/delete": {
            "post": {
                "tags": [
                    "ADMINISTRATION__ADMINISTRATEURS"
                ],
                "summary": "Supprimer un administrateur",
                "operationId": "5cc284c12c57d8844aef11393215dae8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/admins/theme/{id}": {
            "get": {
                "tags": [
                    "ADMINISTRATION__ADMINISTRATEURS"
                ],
                "summary": "Change le thème du design d'un administrateur par son ID",
                "operationId": "77418f766cb0fd05442ce1d2cf9d40d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 18
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "bearerFormat": "JWT"
            },
            "0": []
        }
    }
}